LayoutTests/http/tests/security/contentSecurityPolicy/upgrade-insecure-requests/proper-uir-on-navigation.html

<!DOCTYPE html>
<html>
<head>
    <meta http-equiv="Content-Security-Policy" content="upgrade-insecure-requests">
</head>
<body>
    <script>
    if (window.testRunner)
        testRunner.dumpAsText();
    </script>
    <div>
        <p>The following script is loaded using 'http', but it should be upgraded to 'https' due to the
'upgrade-insecure-requests' header.</p>
        <script src="http://127.0.0.1:8443/security/contentSecurityPolicy/resources/alert-pass.js"></script>
    <div>
    <script>
    if (window.testRunner)
        testRunner.queueLoad("http://127.0.0.1:8000/security/contentSecurityPolicy/upgrade-insecure-requests/resources/insecure-site.html");
    </script>
</body>
</html>
